+++ Processor/CategoryProcessor.php
@@ -8,6 +8,7 @@ use Pim\Bundle\MagentoConnectorBundle\Guesser\WebserviceGuesser;
 use Pim\Bundle\MagentoConnectorBundle\Manager\LocaleManager;
 use Pim\Bundle\MagentoConnectorBundle\Merger\MagentoMappingMerger;
 use Pim\Bundle\MagentoConnectorBundle\Normalizer\AbstractNormalizer;
+use Pim\Bundle\MagentoConnectorBundle\Normalizer\CategoryNormalizer;
 use Pim\Bundle\MagentoConnectorBundle\Normalizer\Exception\NormalizeException;
 use Pim\Bundle\MagentoConnectorBundle\Webservice\MagentoSoapClientParametersRegistry;
 use Pim\Bundle\MagentoConnectorBundle\Normalizer\Exception\CategoryNotMappedException;
@@ -30,8 +31,23 @@ class CategoryProcessor extends AbstractProcessor
     protected $categoryMappingMerger;
 
     /**
+     * @var CategoryNormalizer
+     */
+    protected $categoryNormalizer;
+
+    /**
+     * @var boolean
+     */
+    protected $isAnchor;
+
+    /**
+     * @var boolean
+     */
+    protected $urlKey;
+
+    /**
      * @param WebserviceGuesser                   $webserviceGuesser
-     * @param ProductNormalizerGuesser            $normalizerGuesser
+     * @param NormalizerGuesser                   $normalizerGuesser
      * @param LocaleManager                       $localeManager
      * @param MagentoMappingMerger                $storeViewMappingMerger
      * @param MagentoMappingMerger                $categoryMappingMerger
@@ -89,6 +105,50 @@ class CategoryProcessor extends AbstractProcessor
     }
 
     /**
+     * Get isAnchor.
+     *
+     * @return boolean
+     */
+    public function isIsAnchor()
+    {
+        return $this->isAnchor;
+    }
+
+    /**
+     * Set isAnchor.
+     *
+     * @param boolean $isAnchor
+     *
+     * @return CategoryProcessor
+     */
+    public function setIsAnchor($isAnchor)
+    {
+        $this->isAnchor = $isAnchor;
+
+        return $this;
+    }
+
+    /**
+     * @return boolean
+     */
+    public function isUrlKey()
+    {
+        return $this->urlKey;
+    }
+
+    /**
+     * @param boolean $urlKey
+     *
+     * @return CategoryProcessor
+     */
+    public function setUrlKey($urlKey)
+    {
+        $this->urlKey = $urlKey;
+
+        return $this;
+    }
+
+    /**
      * Function called before all process
      */
     protected function beforeExecute()
@@ -108,7 +168,9 @@ class CategoryProcessor extends AbstractProcessor
                 'defaultLocale'     => $this->defaultLocale,
                 'magentoStoreViews' => $magentoStoreViews,
                 'categoryMapping'   => $this->categoryMappingMerger->getMapping(),
-                'defaultStoreView'  => $this->getDefaultStoreView()
+                'defaultStoreView'  => $this->getDefaultStoreView(),
+                'is_anchor'         => $this->isAnchor,
+                'urlKey'            => $this->urlKey
             ]
         );
     }
@@ -185,6 +247,22 @@ class CategoryProcessor extends AbstractProcessor
     {
         return array_merge(
             parent::getConfigurationFields(),
+            [
+                'isAnchor' => [
+                    'type'    => 'checkbox',
+                    'options' => [
+                        'help'  => 'pim_magento_connector.export.isAnchor.help',
+                        'label' => 'pim_magento_connector.export.isAnchor.label'
+                    ]
+                ],
+                'urlKey' => [
+                    'type'    => 'checkbox',
+                    'options' => [
+                        'help'  => 'pim_magento_connector.export.urlKey.help',
+                        'label' => 'pim_magento_connector.export.urlKey.label'
+                    ]
+                ]
+            ],
             $this->categoryMappingMerger->getConfigurationField()
         );
     }